Replace `UPSERT` with `UPDATE` + `INSERT`
I ran into some memory issues with a large database on a small
Digital Ocean droplet. Trying this approach to see if it works
better.

A full `UPSERT` is not really needed for this table, where the
overwhelming majority of rows will be updates in most syncs,
and the `INSERT` covers missing cases. Overlap will never be an
issue.